eshop-cache-ha/src/test/java/cn/mrcode/cachepdp/eshop/cache/ha/CommandUsingRequestCache.java
package cn.mrcode.cachepdp.eshop.cache.ha;

import com.netflix.hystrix.HystrixCommand;
import com.netflix.hystrix.HystrixCommandGroupKey;

/**
 * ${todo}
 *
 * @author : zhuqiang
 * @date : 2019/6/3 21:51
 */
public class CommandUsingRequestCache extends HystrixCommand<Boolean> {

    private final int value;

    protected CommandUsingRequestCache(int value) {
        super(HystrixCommandGroupKey.Factory.asKey("ExampleGroup"));
        this.value = value;
    }

    @Override
    protected Boolean run() {
        // 当值为 0 或者是 2 的整倍数的时候，返回 true
        System.out.println("run 方法被执行");
        return value == 0 || value % 2 == 0;
    }

    @Override
    protected String getCacheKey() {
        return String.valueOf(value);
    }
}